requirements.txt
python
streamlit
Pillow
torch
torchvision